requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
bcrypt=4.1.2=pypi_0
ca-certificates=2023.12.12=haa95532_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
exceptiongroup=1.2.0=pypi_0
fastapi=0.110.0=pypi_0
h11=0.14.0=pypi_0
idna=3.6=pypi_0
mysql-connector-python=8.3.0=pypi_0
openssl=3.0.13=h2bbff1b_0
pillow=10.2.0=pypi_0
pip=23.3.1=py39haa95532_0
pydantic=2.6.2=pypi_0
pydantic-core=2.16.3=pypi_0
python=3.9.18=h1aa4202_0
python-multipart=0.0.9=pypi_0
setuptools=68.2.2=py39haa95532_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
starlette=0.36.3=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2024a=h04d1e81_0
uvicorn=0.27.1=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.41.2=py39haa95532_0
